Use round() instead of roundf() for CGFloat precision
CGFloat is double on 64-bit macOS. roundf() truncates to float precision.

Related to #309

+++ MacDown/Code/View/MPDocumentSplitView.m
@@ -103,7 +103,7 @@
     CGFloat dividerThickness = self.dividerThickness;
     CGFloat newTotalWidth = self.frame.size.width - dividerThickness;
     CGFloat newHeight = self.frame.size.height;
-    CGFloat newLeftWidth = roundf(newTotalWidth * ratio);
+    CGFloat newLeftWidth = round(newTotalWidth * ratio);
     CGFloat newRightWidth = newTotalWidth - newLeftWidth;
 
     NSView *left = parts[0];
